+++ src/pumpwood_flaskviews/action/__init__.py
@@ -1,0 +1,8 @@
+"""Module to define actions on Pumpwood."""
+from .action import Action, action
+from .load_parameters import LoadActionParameters
+
+
+__all__ = [
+    Action, action, LoadActionParameters
+]