+++ src/pumpwood_flaskviews/model/general.py
@@ -1,7 +1,6 @@
 """Functions and classes for flask/SQLAlchemy models."""
-from loguru import logger
-from flask import g
 from typing import Literal
+from dataclasses import dataclass
 from sqlalchemy.orm import DeclarativeBase
 from sqlalchemy import Column, BigInteger
 from flask_sqlalchemy.query import Query
@@ -9,9 +8,11 @@ from pumpwood_flaskviews.query import (
     BaseQueryABC, BaseQueryNoFilter, SqlalchemyQueryMisc)
 from pumpwood_flaskviews.auth import AuthFactory
 from pumpwood_communication.serializers import CompositePkBase64Converter
-from pumpwood_communication.exceptions import PumpWoodObjectDoesNotExist
+from pumpwood_communication.exceptions import (
+    PumpWoodObjectDoesNotExist, PumpWoodOtherException)
 from pumpwood_communication.cache import default_cache
-# from pumpwood_flaskviews.sqlalchemy import get_session
+from pumpwood_communication.type import PumpwoodDataclassMixin
+from pumpwood_flaskviews.cache import PumpwoodFlaskGCache
 
 
 def _try_convert_int(value: str):
@@ -20,6 +21,22 @@ def _try_convert_int(value: str):
         return int(value)
     except (ValueError, TypeError):
         return value
+
+
+@dataclass
+class FlaskPumpWoodBaseModelCacheHash(PumpwoodDataclassMixin):
+    """Dictionary to create cache hash dict for LocalForeignKeyField."""
+
+    authorization_token: str
+    """Request authorization token."""
+    model_class: str
+    """Model class for the autofill field."""
+    object_pk: str | int | dict
+    """Pk associated with objecto to get the autofill field data."""
+    get_type: Literal['default', 'query']
+    """Field to extract data to fill object."""
+    context: str = 'pumpwood-flaskviews-local-foreignkey-field'
+    """Content of the file that will be returned at the action."""
 
 
 class FlaskPumpWoodBaseModel(DeclarativeBase):
@@ -46,6 +63,12 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
        that tables with more the one partition at least the first one must
        be specified on the queries."""
 
+    HASH_DICT = {
+        'context': 'flaskviews--model-query-retrieve',
+        'model_class': None, 'pk': None, 'get-type': None
+    }
+    """Template for hash dictonary."""
+
     @classmethod
     def build_get_cache_hash(cls, pk: str | int,
                              get_type: Literal['default', 'query']
@@ -56,7 +79,9 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
              pk (str | int):
                 Primary to fetch the data.
              get_type (Literal('default', 'query')):
-                Type of the request that is been made.
+                Type of the request that is been made. 'default' is cache
+                associated with user row_permission filter and 'query'
+                are for caches without the row_permission filter.
 
         Returns:
             Return a dictionary with hash dict to cache get.
@@ -64,8 +89,7 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
         hash_dict = AuthFactory.get_auth_header()
         hash_dict = {
             'context': 'flaskviews--model-query-retrieve',
-            'model_class': cls.__name__,
-            'pk': pk,
+            'model_class': cls.__name__, 'pk': pk,
             'get-type': get_type}
         return default_cache._generate_hash(hash_dict)
 
@@ -145,7 +169,7 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
             return query_result.limit(limit)
 
     @classmethod
-    def default_query_get(cls, pk: str | int,
+    def default_query_get(cls, pk: str | int | dict,
                           base_query: Query = None,
                           raise_error: bool = True) -> object:
         """Get model_class object using pumpwood pk.
@@ -155,8 +179,10 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
         query.get to treat both possibilities.
 
         Args:
-            pk (str, int):
-                Pumpwood primary key.
+            pk (str, int, dict):
+                Pumpwood primary key. If the pk is already a dictonary
+                it will be considered ready to be passed to the
+                query.
             base_query (Query):
                 A base query to be used as initial filter.
             raise_error (bool):
@@ -168,38 +194,53 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
         Return:
             Returns a SQLAlchemy object with corresponding primary key.
         """
-        hash_str = cls.build_get_cache_hash(
-            pk=pk, get_type='default')
-        cache_data = getattr(g, hash_str, None)
+        # Is is not possible to unify the implementation because the cache
+        # for default query uses the base query filter and the query do not.
+        # Unify leads to cache inconstency.
+        hash_dict = FlaskPumpWoodBaseModelCacheHash(
+            authorization_token=AuthFactory.get_auth_header()['Authorization'],
+            model_class=cls.__name__, object_pk=pk,
+            get_type='default')
+        cache_data = PumpwoodFlaskGCache.get(hash_dict=hash_dict)
         if cache_data is not None:
-            logger.info("default_query_get data retrieved from g object")
             return cache_data
 
-        converted_pk = CompositePkBase64Converter.load(pk)
-        if isinstance(converted_pk, (int, float)):
-            # If a numeric data is passed as pk it is associated with
-            # 'id' field, it is necessary to convert to a dict to unpack
-            # on filter_by
-            converted_pk = {'id': converted_pk}
+        converted_pk = None
+        if isinstance(pk, dict):
+            converted_pk = pk
+        else:
+            converted_pk = CompositePkBase64Converter.load(pk)
+            if isinstance(converted_pk, (int, float)):
+                # If a numeric data is passed as pk it is associated with
+                # 'id' field, it is necessary to convert to a dict to unpack
+                # on filter_by
+                converted_pk = {'id': converted_pk}
 
-        # Use base query to filter object acording to user's permission
-        tmp_base_query = cls.default_filter_query(query=base_query)
-
-        # Since base query inject a filter retricting user information
-        # it is not possible to use .get
-        tmp_base_query_2 = tmp_base_query\
-            .filter_by(**converted_pk)
-        model_object = tmp_base_query_2.first()
-
-        if model_object is None and raise_error:
+        # Use base query to filter object acording to user's permission,
+        # it is necessary to use filter_by on request because it is
+        # applied over a previous id
+        model_object_results = cls.default_filter_query(query=base_query)\
+            .filter_by(**converted_pk).all()
+        if len(model_object_results) == 0 and raise_error:
             message = "Requested object {model_class}[{pk}] not found."
             raise PumpWoodObjectDoesNotExist(
                 message=message, payload={
                     "model_class": cls.__name__,
                     "pk": _try_convert_int(pk)})
+        elif len(model_object_results) != 1:
+            msg = (
+                "Get query for {model_class}[{pk}] returned more than "
+                "one object, check implementation.")
+            raise PumpWoodOtherException(
+                message=msg, payload={
+                    "model_class": cls.__name__,
+                    "pk": _try_convert_int(pk)})
+
+        # Get first element
+        model_object = model_object_results[0]
 
         # Set a local cache for object using g object
-        setattr(g, hash_str, model_object)
+        PumpwoodFlaskGCache.set(hash_dict=hash_dict, value=model_object)
         return model_object
 
     @classmethod
@@ -246,7 +287,7 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
             return query_result.limit(limit)
 
     @classmethod
-    def query_get(cls, pk: str | int, base_query: Query = None,
+    def query_get(cls, pk: str | int | dict, base_query: Query = None,
                   raise_error: bool = True) -> object:
         """Get model_class object using pumpwood pk without base query filter.
 
@@ -255,8 +296,10 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
         query.get to treat both possibilities.
 
         Args:
-            pk (str, int):
-                Pumpwood primary key.
+            pk (str, int, dict):
+                Pumpwood primary key. If the pk is already a dictonary
+                it will be considered ready to be passed to the
+                query.
             base_query (Query):
                 A base query to be used as initial filter.
             raise_error (bool):
@@ -265,35 +308,49 @@ class FlaskPumpWoodBaseModel(DeclarativeBase):
         Return:
             Returns a SQLAlchemy object with corresponding primary key.
         """
-        hash_str = cls.build_get_cache_hash(
-            pk=pk, get_type='query')
-        cache_data = getattr(g, hash_str, None)
+        hash_dict = FlaskPumpWoodBaseModelCacheHash(
+            authorization_token=AuthFactory.get_auth_header()['Authorization'],
+            model_class=cls.__name__, object_pk=pk,
+            get_type='query')
+        cache_data = PumpwoodFlaskGCache.get(hash_dict=hash_dict)
         if cache_data is not None:
-            logger.info("default_query_get data retrieved from g object")
             return cache_data
 
-        converted_pk = CompositePkBase64Converter.load(pk)
-        if isinstance(converted_pk, (int, float)):
-            # If a numeric data is passed as pk it is associated with
-            # 'id' field, it is necessary to convert to a dict to unpack
-            # on filter_by
-            converted_pk = {'id': converted_pk}
+        converted_pk = None
+        if isinstance(pk, dict):
+            converted_pk = pk
+        else:
+            converted_pk = CompositePkBase64Converter.load(pk)
+            if isinstance(converted_pk, (int, float)):
+                # If a numeric data is passed as pk it is associated with
+                # 'id' field, it is necessary to convert to a dict to unpack
+                # on filter_by
+                converted_pk = {'id': converted_pk}
 
-        # Use base query to filter object acording to user's permission
+        # Use base query if passed as parameter
         tmp_base_query = cls.query if base_query is None else base_query
 
-        # Since base query inject a filter retricting user information
-        # it is not possible to use .get
-        model_object = tmp_base_query\
-            .filter_by(**converted_pk).first()
-
-        if model_object is None and raise_error:
+        # Filter the objects acording to primary argument and treat the
+        # cases when the object is not found or when the query returns more
+        # then one result
+        model_object_results = tmp_base_query\
+            .filter_by(**converted_pk).all()
+        if len(model_object_results) == 0 and raise_error:
             message = "Requested object {model_class}[{pk}] not found."
             raise PumpWoodObjectDoesNotExist(
                 message=message, payload={
                     "model_class": cls.__name__,
                     "pk": _try_convert_int(pk)})
+        elif len(model_object_results) != 1:
+            msg = (
+                "Get query for {model_class}[{pk}] returned more than "
+                "one object, check implementation.")
+            raise PumpWoodOtherException(
+                message=msg, payload={
+                    "model_class": cls.__name__,
+                    "pk": _try_convert_int(pk)})
+        model_object = model_object_results[0]
 
         # Set a local cache for object using g object
-        setattr(g, hash_str, model_object)
+        PumpwoodFlaskGCache.set(hash_dict=hash_dict, value=model_object)
         return model_object