+++ src/pumpwood_flaskviews/sqlalchemy/connection.py
@@ -38,13 +38,8 @@ class PumpwoodDBGuard:
         """Check if database is avaiable."""
         self.db.session.execute(text("SELECT 1;"))
 
-    def _regenerate_connections(self) -> None:
-        """Regenerate application connections."""
-        self.db.engine.dispose()
-        self.db.session.rollback()
-
-    def before_request_hook(self) -> None:
-        """Function ran before each request."""
+    def check_reset_db(self) -> None:
+        """Check if database is avaiable, reset if not."""
         retry_counter = 0
         max_retries = 5
 
@@ -65,10 +60,19 @@ class PumpwoodDBGuard:
 
                     # Wait a litte before regen the connections
                     time.sleep(0.10)
-                    self._regenerate_connections()
+                    self.regenerate_connections()
                 else:
                     logger.exception("Database unavailable after max retries.")
                     raise e
+
+    def regenerate_connections(self) -> None:
+        """Regenerate application connections."""
+        self.db.session.rollback()
+        self.db.engine.dispose()
+
+    def before_request_hook(self) -> None:
+        """Function ran before each request."""
+        self.check_reset_db()
 
     def teardown_request_hook(self, exception=None) -> None:
         """Function to run after the request."""